Fix for 1 bug and 1 issue in FilterInitMiddleware.  Plus a unit test!
The issue is that we no longer get a Dataset_File save when we've
successfully uploaded the file content (and verified).  That is a save
on the Replica.  So we weren't running the filters when the file content
became available.  This fix adds a 'post_save' signal on Replica which
runs the filter on the Dataset_File if the Replica is now verified.

The bug is a minor one in the way that filter parameters were configured
in the POST_SAVE_FILTERS setting.  If you had both **args and **kwargs, the
former were silently ignored.