Implement comma separated ParameterName choices
ParameterNames have a choices field that stores a string, presumably to list the available choices for a particular parameter name. For the most part this seems to be non-functional except for a breif reference to in DataFile searching. This commit implements logic to extract comma separated choices from the choices string. It also implements logic for providing drop down boxes for choosing available choices in the add and edit metadata view.