This addresses https://github.com/mytardis/mytardis/issues/379
Background: The UserProfile class is primarily used to add an additional field
("isDjangoAccount") to the built-in django.contrib.auth.models.User model. In
many places in MyTardis's code, a UserProfile record is created immediately
after creating a User record (and they are associated as a one-to-one
relationship).  However if a User record is created from the Django Admin
interface, then a UserProfile record won't automatically be created, which can
lead to internal server errors when trying to log in with that User account.

Instead of having to create a UserProfile manually each time a User record is
created, this code automatically creates a UserProfile record (if it doesn't
already exist) when a User record is saved, using a post_save signal hooked up
to the new tardis.tardis_portal.filters.user.UserFilter class.  The existing
configure_user method in tardis/tardis_portal/auth/utils.py is used to create
the UserProfile, and it also adds the user to the default groups for new users
(if any), defined in settings.py as NEW_USER_INITIAL_GROUPS.

The POST_SAVE_FILTERS defined in settings.py are only for post_save signals
associated with DataFile records.
tardis.tardis_portal.filters.FilterInitMiddleware has been modified so that it
can hook up post-save filters not just for DataFile records, but also for User
records.  FilterInitMiddleware's __init__ method is called twice from
tardis/tardis_portal/tasks.py, once for DataFile records and once for User
records.  The two calls to FilterInitMiddleware(...) can't be within the same
try block, because the __init__ method raises MiddlewareNotUsed (for
historical reasons?)