Adding support for querying ObjectACLs with the RESTful API, as required by MyData.
When creating an experiment, MyData initially creates an ObjectACL for the
facility manager (whose MyTardis username is entered into MyData's Settings
dialog).  The facility manager is the initial owner of the created experiment
in MyTardis.  Then if the data owner (a user or group, usually inferred by a
folder name, indicating the researcher(s) who collected the data) is different
from the facility manager, then MyData will create a second ObjectACL record,
granting access to the experiment to the data owner / researcher(s).