Fix adding Users and Groups via the UI by removing calls to transaction.commit().
This is probably a bug carried over after the port to Django 1.6 - there may be other
The @transaction.atomic decorator is in use on these functions, so transaction.commit() fails.
I've also wrapped some parts in additional try/except: transaction.rollback() clauses.
There are additional deprecated instances of transaction.commit_* in the code that are not yet addressed.