Moved StorageBoxResourceClass above StorageBoxOptionResource and StorageBoxAttributeResource classes.
Restricted querying of StorageBoxOptions to options whose key is listed in
StorageBoxResource.accessible_keys.

For direct querying of StorageBoxOptions, the restriction is done in the
ACLAuthorization class.

But for querying the StorageBoxResource (which has a StorageBoxOption
ToManyField with full=True), we use a trick found in the TastyPie docs
allowing ToManyField's attribute argument to use a callable:
http://django-tastypie.readthedocs.org/en/latest/fields.html#tomanyfield