Addressing angular/controller-as (reported by ESLint). See: https://github.com/Gillespie59/eslint-plugin-angular/blob/master/docs/controller-as.md https://github.com/johnpapa/angular-styleguide/blob/master/a1/README.md#style-y031
Also removed global variables, which I wanted to do anyway, and it
seemed to be necessary when switching to the controllerAs syntax.

Instead of defining experiment_id, is_publication and
is_publication_draft as globals in the view_experiment.html template,
they are passed into the PublicationFormController as attributes of
the controller's div element when instantiating the experiment view
page's PublicationFormController.  Then when instantiating the
ngDialog's PublicationFormController, we use the ngDialog.open's "data"
parameter to pass experiment_id, is_publication and is_publication_draft
to the ngDialog's controller.

Instead of using a global publication_id variable to save the result
from the ngDialog's PublicationFormController and make it available to
the experiment view's PublicationFormController, we pass it as an
argument in ngDialog's closeThisDialog, and then receive it as an
argument in preCloseCallback.