Removing dataset.is_online check because it contributes significantly to slow page loading for large datasets, as discussed here: https://github.com/mytardis/mytardis/issues/595
Instead of trying to check the status of every file in the dataset
when the page loads, we can have the ajax check only the current
page of displayed files.

Eventually we want a better user experience for tape recalls, but
the "Some or all files in this dataset are archived" warning resulting
from the dataset.is_online check has not proved useful enough to
justify the slow page loading time.

As noted in the linked GitHub issue, it may be possible to
do the dataset.is_online query more efficiently with aggregate /
group by queries, but personally I don't think it's worth the
trouble.

It would be possible to write an asynchronous task to check datasets
periodically and add some metadata indicating whether their files
are all available on disk, or whether some of all are on tape, but
that is beyond the scope of this pull request.