Marking deprecated functionality.
I have marked it as being deprecated in v3.10, which would mean
that the deprecation warnings should appear in v3.8, but it's
up to @keithschulze (as our v3.8 release manager) to decide whether
these deprecation warnings can be included in v3.8, given that
we're in a feature freeze.  If not, I'll update the PR to indicate
that the deprecated code will be removed in v3.11 (not v3.10).