Removed import via staging functionality which used jsTree.
This code has been broken for a while.  Developers are welcome to
submit pull requests adding similar functionality back into a MyTardis
add-on app, but fixing this old code is not a priority for MyTardis
core developers at the moment, giving that it is not used in any known
deployments maintained by MyTardis's developers.