Improving downloads and recall requests for S3 and HSM files.
For S3 objects, it is better to redirect the download request to
the Object Store, rather than streaming it via Django.

For files on Hierarchical Storage Management (HSM) systems, we
need to check if they are offline (zero blocks on disk) before
attempting to download them, and we need to offer the user the
ability to request recalling a file from tape / archive.